# src/TicketManagement.WebUI/OpenAPIs/swagger.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "My Title",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/AreaManagement/area/{areaId}": {
      "delete": {
        "tags": [
          "AreaManagement"
        ],
        "summary": "Delete selected area.",
        "operationId": "AreaManagement_DeleteArea",
        "parameters": [
          {
            "name": "areaId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "."
          }
        }
      },
      "get": {
        "tags": [
          "AreaManagement"
        ],
        "summary": "Returns selected area.",
        "operationId": "AreaManagement_GetByIdArea",
        "parameters": [
          {
            "name": "areaId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": ".",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Area"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/EventAreaManagement/eventArea": {
      "post": {
        "tags": [
          "EventAreaManagement"
        ],
        "summary": "Add new eventArea.",
        "operationId": "EventAreaManagement_InsertEventArea",
        "requestBody": {
          "x-name": "eventArea",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventArea"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "."
          }
        }
      },
      "put": {
        "tags": [
          "EventAreaManagement"
        ],
        "summary": "Update selected eventArea.",
        "operationId": "EventAreaManagement_UpdateEventArea",
        "requestBody": {
          "x-name": "eventArea",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventArea"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "."
          }
        }
      }
    },
    "/api/EventAreaManagement/eventArea/{eventAreaId}": {
      "delete": {
        "tags": [
          "EventAreaManagement"
        ],
        "summary": "Delete selected eventArea.",
        "operationId": "EventAreaManagement_DeleteEventArea",
        "parameters": [
          {
            "name": "eventAreaId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "."
          }
        }
      },
      "get": {
        "tags": [
          "EventAreaManagement"
        ],
        "summary": "Returns selected eventArea.",
        "operationId": "EventAreaManagement_GetByIdEventArea",
        "parameters": [
          {
            "name": "eventAreaId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": ".",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventArea"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/EventAreaManagement/EventAreasByEventId/{eventId}": {
      "get": {
        "tags": [
          "EventAreaManagement"
        ],
        "summary": "Returns list of the eventAreas into selected eventId.",
        "operationId": "EventAreaManagement_GetAllEventAreasByEventId",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": ".",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventArea"
                  }
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/EventManagement/event/{eventId}": {
      "delete": {
        "tags": [
          "EventManagement"
        ],
        "summary": "Delete selected event.",
        "operationId": "EventManagement_DeleteEvent",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "."
          }
        }
      },
      "get": {
        "tags": [
          "EventManagement"
        ],
        "summary": "Returns selected event.",
        "operationId": "EventManagement_GetByIdEvent",
        "parameters": [
          {
            "name": "eventId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": ".",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/SeatManagement/seat": {
      "post": {
        "tags": [
          "SeatManagement"
        ],
        "summary": "Add new seat.",
        "operationId": "SeatManagement_InsertSeat",
        "requestBody": {
          "x-name": "seat",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Seat"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "."
          }
        }
      },
      "put": {
        "tags": [
          "SeatManagement"
        ],
        "summary": "Update selected seat.",
        "operationId": "SeatManagement_UpdateSeat",
        "requestBody": {
          "x-name": "seat",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Seat"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "."
          }
        }
      }
    },
    "/api/SeatManagement/seat/{seatId}": {
      "delete": {
        "tags": [
          "SeatManagement"
        ],
        "summary": "Delete selected seat.",
        "operationId": "SeatManagement_DeleteSeat",
        "parameters": [
          {
            "name": "seatId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "."
          }
        }
      },
      "get": {
        "tags": [
          "SeatManagement"
        ],
        "summary": "Returns selected seat.",
        "operationId": "SeatManagement_GetByIdSeat",
        "parameters": [
          {
            "name": "seatId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": ".",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Seat"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/SeatManagement/SeatsByAreaId/{areaId}": {
      "get": {
        "tags": [
          "SeatManagement"
        ],
        "summary": "Returns list of the seats into selected areaId.",
        "operationId": "SeatManagement_GetAllSeatsByAreaId",
        "parameters": [
          {
            "name": "areaId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": ".",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Seat"
                  }
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Area": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseEntity"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "layoutId",
              "description",
              "coordX",
              "coordY"
            ],
            "properties": {
              "layoutId": {
                "type": "integer",
                "format": "int32"
              },
              "description": {
                "type": "string",
                "maxLength": 200,
                "minLength": 1
              },
              "coordX": {
                "type": "integer",
                "format": "int32"
              },
              "coordY": {
                "type": "integer",
                "format": "int32"
              }
            }
          }
        ]
      },
      "BaseEntity": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "ProblemDetails": {
        "type": "object",
        "additionalProperties": {
          "nullable": true
        },
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "EventArea": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseEntity"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "eventId",
              "description",
              "coordX",
              "coordY",
              "price"
            ],
            "properties": {
              "eventId": {
                "type": "integer",
                "format": "int32"
              },
              "description": {
                "type": "string",
                "maxLength": 200,
                "minLength": 1
              },
              "coordX": {
                "type": "integer",
                "format": "int32"
              },
              "coordY": {
                "type": "integer",
                "format": "int32"
              },
              "price": {
                "title": "Price",
                "type": "number",
                "format": "decimal"
              }
            }
          }
        ]
      },
      "Event": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseEntity"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "name",
              "eventTime",
              "description",
              "layoutId",
              "eventEndTime",
              "eventLogoImage"
            ],
            "properties": {
              "name": {
                "type": "string",
                "maxLength": 120,
                "minLength": 1
              },
              "eventTime": {
                "type": "string",
                "format": "date-time",
                "minLength": 1
              },
              "description": {
                "type": "string",
                "minLength": 1
              },
              "layoutId": {
                "type": "integer",
                "format": "int32"
              },
              "eventEndTime": {
                "type": "string",
                "format": "date-time",
                "minLength": 1
              },
              "eventLogoImage": {
                "type": "string",
                "minLength": 1
              }
            }
          }
        ]
      },
...
      "Seat": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseEntity"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "areaId",
              "row",
              "number"
            ],
            "properties": {
              "areaId": {
                "type": "integer",
                "format": "int32"
              },
              "row": {
                "type": "integer",
                "format": "int32"
              },
              "number": {
                "type": "integer",
                "format": "int32"
              }
            }
          }
        ]
      },
...
    }
  }
}